Organize zustand stores inside one folder
