+++ client/src/components/NotificationModal/NotificationModal.jsx
@@ -27,7 +27,7 @@
 
 import React from 'react';
 import { useNavigate } from 'react-router-dom';
-import useStore from '../../useStore';
+import useStore from '../../store/useStore';
 import { styled } from '@mui/material/styles';
 import Dialog from '@mui/material/Dialog';
 import DialogTitle from '@mui/material/DialogTitle';
