+++ client/src/store/unitsStore.js
@@ -3,7 +3,7 @@
 
   Import:
 
-    import useUnitsStore from '../../../unitsStore';
+    import useUnitsStore from '../../../store/unitsStore';
 
   Get checked units:
